% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachment_format.R
\name{attachment_format}
\alias{attachment_format}
\title{Pack Plot, Text, and Color to attachment format}
\usage{
attachment_format(former_text, status_color, image_url = NULL,
  image_title = " ")
}
\arguments{
\item{former_text}{your text on attachment}

\item{status_color}{hex}

\item{image_url}{(optional) your plot url, must be https}

\item{image_title}{(optional) text}
}
\value{
json file to attach
}
\description{
Pack Plot, Text, and Color to attachment format
}
\examples{
attachment_format('Order Status: Over High', '#FF0000', 'https://image.url', '20190901_Order')
}
