% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot_define.R
\name{box_plot_define}
\alias{box_plot_define}
\title{Define outlier by quantile}
\usage{
box_plot_define(boxplot_list, number)
}
\arguments{
\item{boxplot_list}{list, generate with boxplot()}

\item{number, }{numeric}
}
\value{
list with status and color
}
\description{
Define outlier by quantile
}
\examples{
\dontrun{
    quantile_list <- boxplot(orders$count)
    box_plot_define(quantile_list, 38)
    // will return c('normal', '#31a7ff')
}
}
