% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upload.R
\name{plot_upload}
\alias{plot_upload}
\title{Post ggPlot to slack channel and return private URL}
\usage{
plot_upload(plot = last_plot(), file = "ggplot", dpi = 300,
  scale = 1, slack_api_token = Sys.getenv("SLACK_UPLOAD_API_TOKEN"),
  upload_channel_name = Sys.getenv("SLACK_UPLOAD_IMAGE_CHANNEL"))
}
\arguments{
\item{plot}{ggplot object, default last plot}

\item{file}{prefix of your plot, default ggplot}

\item{dpi}{default 300}

\item{scale}{default 1}

\item{slack_api_token}{file:upload token, or setup with slackeonnier_setup()}

\item{upload_channel_name}{string, default #general}
}
\value{
private url on slack
}
\description{
Post ggPlot to slack channel and return private URL
}
\examples{
plot_upload(my.plot)
}
