% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackeonnier_sent.R
\name{slackeonnier_sent}
\alias{slackeonnier_sent}
\title{Send attachment to Slack}
\usage{
slackeonnier_sent(attachment,
  webhook = Sys.getenv("SLACK_INCOMING_WEBHOOK_URL"),
  channel = Sys.getenv("SLACK_SENT_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"))
}
\arguments{
\item{attachment}{json, comes from attachment_format()}

\item{webhook}{url, use slackeonnier_setup() to set up}

\item{channel}{string, use slackeonnier_setup() to set up}

\item{username}{string, use slackeonnier_setup() to set up}
}
\description{
Send attachment to Slack
}
