% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackeonnier_setup.R
\name{slackeonnier_setup}
\alias{slackeonnier_setup}
\title{Setup your api token, incoming webhook, channels}
\usage{
slackeonnier_setup(upload_api_token = "",
  upload_image_channel = "#random", incoming_webhook_url = "",
  sent_channel = "#general", username = "slackeonnier")
}
\arguments{
\item{upload_api_token}{file:upload token}

\item{upload_image_channel}{alternative channel for uploading plot, default #random}

\item{incoming_webhook_url}{incoming webhook for posting notification}

\item{sent_channel}{will post notification to this channel, default #general}

\item{username}{name of slack App, default slackeonnier}
}
\description{
Setup your api token, incoming webhook, channels
}
\examples{
\dontrun{
slackeonnier()
}
}
